Initial commit of some kmer utilities.
there are two utilties included.

one is kmer_frequency_per_sequence,
which outputs a (m x n) matrix where m is the sequence, and n is the
frequency of that nmer to occur in the given sequence.

the other tool is kmer_total_count, which counts kmers for the total
file, not just one sequence